backend/requirements.txt
  anthropic==0.40.0
  langgraph==0.2.50                                              
  langchain==0.3.0
  langchain-anthropic==0.3.0                                     
  langchain-community==0.3.0
  tavily-python==0.5.0                           
  fastapi==0.115.0
  uvicorn==0.32.0
  httpx==0.27.0
  python-dotenv==1.0.0
  pydantic==2.9.0
  requests==2.32.0
  beautifulsoup4==4.12.0